WIP (module): refactor the whole thing, add some new props, improve perfs...
This is still a quite early stage, not everything has been tested
properly

* Improve perfs by removing A LOT of useless function binding
* Clean up the code where some fallbacks were useless
* Add missing propTypes validation
* Add renderError prop (to be documented)
* This might fix a bug where the gallery crashed if an item was removed
* The gallery won’t crash if RN cannot fetch the dimensions of a remote
image
* …
